Change calc structure for support BigInt

+++ src/calc.h
@@ -2,4 +2,4 @@
 #include "stack.h"
 #include "bigint.h"
 
-int CalcExp(const CalcData* expr, long long* result);
+int CalcExp(const CalcData* expr, BigInt* result);
